proxy_protocol: pack submission for 64-bit size and offsets
Shrink opcode/flags/channel_id and memview ids so transfer size and
offsets can be full 64-bit values without growing the 64-byte ring record.

+++ src/core/device_proxy/proxy_runtime.h
@@ -177,7 +177,7 @@ private:
     };
 
     struct RegistryEntry {
-        uint64_t proxy_memview_id = 0;
+        uint32_t proxy_memview_id = 0;
         nixlMemViewH backend_memview = nullptr;
         ProxyMemViewRegEntryState state = ProxyMemViewRegEntryState::ENTRY_ALLOCATED;
         ProxyMemViewRegMetadataKind metadata_kind = ProxyMemViewRegMetadataKind::METADATA_KIND_NONE;
